@@ -18,7 +18,7 @@
   "require": {
     "php": "^7.2",
     "psr/log": "^1.1",
-    "symfony/property-access": "^3.4|^4.3|^5",
+    "symfony/property-access": "^4.3|^5",
     "guzzlehttp/guzzle": "^6.0",
     "justinrainbow/json-schema": "^5.0",
     "myclabs/php-enum": "^1.6",
@@ -27,14 +27,14 @@
     "ext-json": "*"
   },
   "require-dev": {
-    "phpunit/phpunit": "^8.0|^9.0",
+    "phpunit/phpunit": "^9.0",
     "symfony/config": "^4.3|^5",
     "symfony/console": "^4.3|^5",
     "doctrine/collections": "^1.6",
     "doctrine/orm": "^2.6",
     "enqueue/amqp-bunny": "^0.10",
     "enqueue/enqueue": "^0.10",
-    "abc/scheduler": "dev-master",
+    "abc/scheduler": "^0.2",
     "zircote/swagger-php": "^3.0"
   },
   "autoload": {
