@@ -4,6 +4,7 @@ namespace Abc\Job\Schedule;
 
 use Abc\Job\CronJobManager;
 use Abc\Job\Model\CronJobInterface;
+use Abc\Job\Model\JobManagerInterface;
 use Abc\Scheduler\ProviderInterface;
 use Abc\Scheduler\ScheduleInterface as BaseScheduleInterface;
 
@@ -12,11 +13,17 @@ class ScheduleProvider implements ProviderInterface
     /**
      * @var CronJobManager
      */
-    private $cronJob;
+    private $cronJobManager;
 
-    public function __construct(CronJobManager $manager)
+    /**
+     * @var JobManagerInterface
+     */
+    private $jobManager;
+
+    public function __construct(CronJobManager $cronJobManager, JobManagerInterface $jobManager)
     {
-        $this->cronJob = $manager;
+        $this->cronJobManager = $cronJobManager;
+        $this->jobManager = $jobManager;
     }
 
     public function getName(): string
@@ -26,17 +33,30 @@ class ScheduleProvider implements ProviderInterface
 
     public function provideSchedules(int $limit = null, int $offset = null): array
     {
-        return $this->cronJob->list(null, null, $limit, $offset);
+        return $this->cronJobManager->list(null, null, $limit, $offset);
+    }
+
+    public function existsConcurrent(BaseScheduleInterface $schedule): bool
+    {
+        $this->assertScheduleType($schedule);
+
+        /** @var CronJobInterface $schedule */
+        return $this->jobManager->existsConcurrent($schedule->getJob());
     }
 
     public function save(BaseScheduleInterface $schedule): void
     {
+        $this->assertScheduleType($schedule);
+
         /** @var CronJobInterface $schedule */
-        if(!$schedule instanceof CronJobInterface)
-        {
+        $this->cronJobManager->update($schedule);
+    }
+
+    private function assertScheduleType(BaseScheduleInterface $schedule): void
+    {
+
+        if (!$schedule instanceof CronJobInterface) {
             throw new \InvalidArgumentException(sprintf('Schedule must implement %s', CronJobInterface::class));
         }
-
-        $this->cronJob->update($schedule);
     }
 }
