@@ -6,18 +6,31 @@ use Abc\Job\Job;
 use Abc\Job\Model\CronJob;
 use Abc\Job\Model\CronJobInterface;
 use Abc\Job\Type;
+use Abc\Scheduler\ConcurrencyPolicy;
 use PHPUnit\Framework\TestCase;
 
 class CronJobTest extends TestCase
 {
+    public function testDefaultConcurrencyPolicy()
+    {
+        $subject = new CronJob(
+            '* * * * *', Job::fromArray(
+            [
+                'type' => Type::JOB(),
+                'name' => 'someName',
+            ]
+        ));
+
+        $this->assertEquals((string)ConcurrencyPolicy::ALLOW(), $subject->getConcurrencyPolicy());
+    }
+
     public function testGetScheduleWithConstructorNotCalled()
     {
         /** @var CronJob $subject */
         $subject = $this->getMockBuilder(CronJob::class)
             ->disableOriginalConstructor()
             ->setMethods()
-            ->getMock()
-        ;
+            ->getMock();
 
         $this->expectException(\LogicException::class);
 
@@ -32,8 +45,7 @@ class CronJobTest extends TestCase
                 'type' => Type::JOB(),
                 'name' => 'someName',
             ]
-        )
-        );
+        ));
 
         $this->assertEquals('', $subject->getSchedule());
     }
@@ -44,8 +56,7 @@ class CronJobTest extends TestCase
         $subject = $this->getMockBuilder(CronJob::class)
             ->disableOriginalConstructor()
             ->setMethods()
-            ->getMock()
-        ;
+            ->getMock();
 
         $job = Job::fromArray(
             [
@@ -67,8 +78,7 @@ class CronJobTest extends TestCase
         $subject = $this->getMockBuilder(CronJob::class)
             ->disableOriginalConstructor()
             ->setMethods()
-            ->getMock()
-        ;
+            ->getMock();
 
         $this->expectException(\LogicException::class);
 
@@ -77,14 +87,12 @@ class CronJobTest extends TestCase
 
     public function testGetJobWithJobSet()
     {
-        $subject = new CronJob(
-            '* * * * *', Job::fromArray(
+        $subject = new CronJob('* * * * *', Job::fromArray(
             [
                 'type' => Type::JOB(),
                 'name' => 'someName',
             ]
-        )
-        );
+        ));
 
         $job = Job::fromArray(
             [
@@ -104,8 +112,7 @@ class CronJobTest extends TestCase
         $subject = $this->getMockBuilder(CronJob::class)
             ->disableOriginalConstructor()
             ->setMethods()
-            ->getMock()
-        ;
+            ->getMock();
 
         $this->assertNull($subject->getJobJson());
     }
@@ -116,8 +123,7 @@ class CronJobTest extends TestCase
         $subject = $this->getMockBuilder(CronJob::class)
             ->disableOriginalConstructor()
             ->setMethods()
-            ->getMock()
-        ;
+            ->getMock();
 
         $subject->setJobJson('someJson');
 
@@ -138,8 +144,7 @@ class CronJobTest extends TestCase
 
         $this->assertInstanceOf(CronJobInterface::class, $cronJob);
         $this->assertSame('* * * * *', $cronJob->getSchedule());
-        $this->assertEquals(
-            Job::fromArray(
+        $this->assertEquals(Job::fromArray(
                 [
                     'type' => Type::JOB(),
                     'name' => 'someName',
@@ -149,8 +154,25 @@ class CronJobTest extends TestCase
         );
     }
 
+    public function testFromJsonWithConcurrencyPolicy()
+    {
+        $cronJob = CronJob::fromJson(
+            json_encode(
+                [
+                    'schedule' => '* * * * *',
+                    'type' => Type::JOB(),
+                    'name' => 'someName',
+                    'concurrencyPolicy' => (string) ConcurrencyPolicy::FORBID()
+                ]
+            )
+        );
+
+        $this->assertInstanceOf(CronJobInterface::class, $cronJob);
+        $this->assertEquals(ConcurrencyPolicy::FORBID(), $cronJob->getConcurrencyPolicy());
+    }
+
     /**
-     * @param string $json
+     * @param  string  $json
      * @dataProvider provideInvalidJson
      */
     public function testFromJsonWithInvalidData(string $json)
@@ -168,6 +190,7 @@ class CronJobTest extends TestCase
                     'schedule' => '* * * * *',
                     'type' => Type::JOB(),
                     'name' => 'someName',
+                    'concurrencyPolicy' => (string) ConcurrencyPolicy::FORBID()
                 ]
             )
         );
@@ -175,16 +198,9 @@ class CronJobTest extends TestCase
         $object = json_decode($cronJob->toJson());
 
         $this->assertEquals($cronJob->getSchedule(), $object->schedule);
-        $this->assertEquals(
-            $cronJob->getJob()
-                ->getType(),
-            $object->type
-        );
-        $this->assertEquals(
-            $cronJob->getJob()
-                ->getName(),
-            $object->name
-        );
+        $this->assertEquals($cronJob->getJob()->getType(), $object->type);
+        $this->assertEquals($cronJob->getJob()->getName(), $object->name);
+        $this->assertEquals((string) $cronJob->getConcurrencyPolicy(), $object->concurrencyPolicy);
     }
 
     public function testFromArray()
@@ -210,6 +226,21 @@ class CronJobTest extends TestCase
         );
     }
 
+    public function testFromArrayWithConcurrencyPolicy()
+    {
+        $cronJob = CronJob::fromArray(
+            [
+                'schedule' => '* * * * *',
+                'concurrencyPolicy' => (string) ConcurrencyPolicy::FORBID(),
+                'type' => Type::JOB(),
+                'name' => 'someName',
+            ]
+        );
+
+        $this->assertInstanceOf(CronJobInterface::class, $cronJob);
+        $this->assertEquals((string) ConcurrencyPolicy::FORBID(), (string) $cronJob->getConcurrencyPolicy());
+    }
+
     public function testToArray()
     {
         $cronJob = CronJob::fromJson(
@@ -225,16 +256,9 @@ class CronJobTest extends TestCase
         $data = $cronJob->toArray();
 
         $this->assertEquals($cronJob->getSchedule(), $data['schedule']);
-        $this->assertEquals(
-            $cronJob->getJob()
-                ->getType(),
-            $data['type']
-        );
-        $this->assertEquals(
-            $cronJob->getJob()
-                ->getName(),
-            $data['name']
-        );
+        $this->assertEquals((string) $cronJob->getConcurrencyPolicy(), $data['concurrencyPolicy']);
+        $this->assertEquals($cronJob->getJob()->getType(), $data['type']);
+        $this->assertEquals($cronJob->getJob()->getName(), $data['name']);
         $this->assertFalse(isset($data['id']));
         $this->assertFalse(isset($data['createdAt']));
         $this->assertFalse(isset($data['updatedAt']));
@@ -259,16 +283,8 @@ class CronJobTest extends TestCase
         $data = $cronJob->toArray();
 
         $this->assertEquals('someId', $data['id']);
-        $this->assertEquals(
-            $cronJob->getCreatedAt()
-                ->format('c'),
-            $data['created']
-        );
-        $this->assertEquals(
-            $cronJob->getUpdatedAt()
-                ->format('c'),
-            $data['updated']
-        );
+        $this->assertEquals($cronJob->getCreatedAt()->format('c'), $data['created']);
+        $this->assertEquals($cronJob->getUpdatedAt()->format('c'), $data['updated']);
     }
 
     public function testToArraySortsKeys()
@@ -293,6 +309,7 @@ class CronJobTest extends TestCase
             [
                 'id',
                 'schedule',
+                'concurrencyPolicy',
                 'type',
                 'name',
                 'input',
@@ -329,7 +346,19 @@ class CronJobTest extends TestCase
         return [
             ['someString'],
             [json_encode(['someString'])],
-            [json_encode(['schedule' => '* * * * *'])],
+            [json_encode(['schedule' => '* * * * *', 'type' => 'Job'])],
+            [json_encode(['name' => 'SomeJobName', 'type' => 'Job'])],
+            [json_encode(['schedule' => '* * * * *', 'name' => 'SomeJobName'])],
+            [
+                json_encode(
+                    [
+                        'schedule' => '* * * * *',
+                        'name' => 'SomeJobName',
+                        'type' => 'Job',
+                        'concurrencyPolicy' => 'invalid'
+                    ]
+                )
+            ]
         ];
     }
 }
