@@ -3,7 +3,9 @@
 namespace Abc\Job\Tests\Schedule;
 
 use Abc\Job\CronJobManager;
+use Abc\Job\Job;
 use Abc\Job\Model\CronJobInterface;
+use Abc\Job\Model\JobManagerInterface;
 use Abc\Job\Schedule\ScheduleProvider;
 use Abc\Scheduler\ScheduleInterface as BaseScheduleInterface;
 use PHPUnit\Framework\MockObject\MockObject;
@@ -17,6 +19,11 @@ class ScheduleProviderTest extends TestCase
     private $cronJobManager;
 
     /**
+     * @var JobManagerInterface|MockObject
+     */
+    private $jobManager;
+
+    /**
      * @var ScheduleProvider
      */
     private $subject;
@@ -24,7 +31,13 @@ class ScheduleProviderTest extends TestCase
     public function setUp(): void
     {
         $this->cronJobManager = $this->createMock(CronJobManager::class);
-        $this->subject = new ScheduleProvider($this->cronJobManager);
+        $this->jobManager = $this->createMock(JobManagerInterface::class);
+        $this->subject = new ScheduleProvider($this->cronJobManager, $this->jobManager);
+    }
+
+    public function testGetName()
+    {
+        $this->assertEquals('abc_job', $this->subject->getName());
     }
 
     public function testProviderSchedules()
@@ -32,6 +45,25 @@ class ScheduleProviderTest extends TestCase
         $this->cronJobManager->expects($this->once())->method('list')->with(null, null, 10, 20)->willReturn(['foobar']);
 
         $this->assertEquals(['foobar'], $this->subject->provideSchedules(10, 20));
+    }
+
+    public function testExistsConcurrent()
+    {
+        $job = $this->createMock(Job::class);
+        $cronJob = $this->createMock(CronJobInterface::class);
+
+        $cronJob->expects($this->once())->method('getJob')->willReturn($job);
+
+        $this->jobManager->expects($this->once())->method('existsConcurrent')->with($job)->willReturn(false);
+
+        $this->assertFalse($this->subject->existsConcurrent($cronJob));
+    }
+
+    public function testExistsConcurrentWithInvalidSchedule()
+    {
+        $this->expectException(\InvalidArgumentException::class);
+
+        $this->subject->existsConcurrent($this->createMock(BaseScheduleInterface::class));
     }
 
     public function testSave()
